requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
groq==0.5.0
python-multipart==0.0.9
